feat: refactoring base for better definition
