+++ base/domain/common/exceptions/__init__.py
