+++ base/domain/common/exceptions/base_domain_exception.py
