+++ base/domain/common/value_objects/__init__.py
