+++ tests/base/exceptions/test_base_domain_exception.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from base.domain.exceptions import BaseDomainException
+from base.domain.common.exceptions import BaseDomainException
 
 
 def test_base_domain_exception_full_validation():
