+++ tests/base/value_objects/test_string_value_object.py
@@ -4,7 +4,7 @@
 import pytest
 
 
-from base.domain.value_objects import StringValueObject
+from base.domain.common.value_objects import StringValueObject
 
 
 def test_string_value_object_validation_pass():
